moved utils into api package moved config into api package updated references
