+++ api/core/account.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/NlaakStudios/Blockchain/config"
+	"github.com/NlaakStudios/Blockchain/api/config"
 	"github.com/boltdb/bolt"
 )
