+++ api/core/wallets.go
@@ -9,7 +9,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/NlaakStudios/Blockchain/config"
+	"github.com/NlaakStudios/Blockchain/api/config"
 )
 
 // Wallets stores a collection of wallets
