--- utils/README.md
@@ -1,39 +1,0 @@
-# utils
---
-    import "\github.com\NlaakStudios\Blockchain\utils"
-
-
-## Usage
-
-#### func  Base58Decode
-
-```go
-func Base58Decode(input []byte) []byte
-```
-Base58Decode decodes Base58-encoded data
-
-#### func  Base58Encode
-
-```go
-func Base58Encode(input []byte) []byte
-```
-Base58Encode encodes a byte array to Base58
-
-#### func  CreateDirIfNotExist
-
-```go
-func CreateDirIfNotExist(dir string)
-```
-
-#### func  IntToHex
-
-```go
-func IntToHex(num int64) []byte
-```
-
-#### func  ReverseBytes
-
-```go
-func ReverseBytes(data []byte)
-```
-ReverseBytes reverses a byte array
